code/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
asyncpg==0.29.0
aioredis==1.3.1
redis==4.6.0
pydantic==2.6.4
python-dotenv==1.0.0
httptools==0.5.0
uvloop==0.17.0
psycopg2-binary==2.9.9
python-multipart==0.0.6